requirements.txt
sqlalchemy
faker>=19.0.0
openai>=1.0.0
python-dotenv
pydantic